migrate autorxs to ts
