+++ src/models/autorxs.ts
@@ -1,0 +1,10 @@
+type Autor = {
+  nombre: string;
+  imagen: string;
+  info: string;
+  enlaceVer: string;
+  notas?: string[];
+  twitter?: string;
+};
+
+export type { Autor };
